Moved table rows to new location for nvi tables
